Fix for thrift server hanging
Summary:
When a socket is initially created and a read callback is set, the `submitNow` flag is being set to false during connection phase. After connection is successfully established, `allowReads()` is called which calls `setReadCB()` with the existing callback. However, in the `ReadSqe::setReadCallback()` method, there was an early exit condition. Since the callback was the same as before and submitNow was set to false, this early exit condition was triggered.

Adding a check for !submitNow and if the request is inflight to the early exit function for fixing the problem.

Adding a new test `ReadCallbackSetDuringConnect` to verify this specific functionality.

Reviewed By: spikeh

Differential Revision: D82485279

fbshipit-source-id: 386d1ceaa684059793a78f99b8cf2fa294f67a26