Instantiate byteEventHelper in AsyncSSLSocket::enableByteEvents() edge case
Summary: AsyncSocket will only call `enableByteEvents()` if the `byteEventHelper_` is `nullptr`, and in its own implementation, `enableByteEvents()` will conditionally instantiate `byteEventHelper_`. However, in AsyncSSLSocket's override, it will call `failByteEvents` in some cases of disabled TLS versions, which `XCHECK`s `byteEventHelper_` is not nullptr, leading to a crash. This diff just instantiates it if it's not already to fix that crash.

Reviewed By: yfeldblum

Differential Revision: D82476941

fbshipit-source-id: a67f50a1ee35a02820fb41b75ef9694862009aaf