Add iteration for Notes in Sections and Segments
Summary:
This patch adds iteration for notes in Segments and Sections, letting a user specify a lambda to search for a note in a given section/segment or all sections or segments.

The logic is effectivelly the same for both, and so I've extracted the actual note parsing logic into it's own private helper function.

Instead of returning the shdr or the phdr, I instead return the note and it's body. I think this the most user friendly API as if we return a Segment, the user would potentially have to reparse all the notes in that PT_NOTE array. However, I think an expected could be an even better API by supporting an error code or message when the note does not exist instead of returning a nullptr.

Reviewed By: yfeldblum

Differential Revision: D82133391

fbshipit-source-id: bdb17816691ba58961f91353fd1e0ac75b027e1a