Remove virtual inheritance from Task/AsyncGenerator/ViaIfAsync
Summary:
## Motivation / historical context

This is a redo of D61401600, which was committed & reverted a long while ago, since it turned out the change (while correct) broke symmetric transfer on Clang 12. This manifested as stack overflow errors in applications that iterated over larger AsyncGenerators, but only in opt builds.

Since then:
 - I improved the symmetric transfer unit test to reliably detect this failure mode.
 - Clang 12 is no longer used by the platform that last saw the problem (see test plan).
 - I have a concrete problem with task wrappers (like `now_task`) that is hard to solve well while promise types are virtual.

So, I'd like to try landing this again, and seeing if it sticks.

The best alternative would probably be a *much* larger refactor of `Task` to allow injecting the behaviors currently provided by wrappers.

## Technical summary

The current virtual interface **only** provides `getErrorHandle`, which can be provided via a manual vtable, instead of an automatic one.

Why do this instead of e.g. marking `getErrorHandle` final on every child?
 - **Mainly:** Wrappers currently don't provide `getErrorHandle`, and cannot do so correctly & performantly due to a few reasons that all converge on the following problem. Adding a virtual `getErrorHandle` -- even if `final` -- bloats the class layout such that the wrapped & original promise pointers are no longer punnable (yes, UB, but nonetheless working UB that nobody noticed up until today), and makes it genuinely hard to obtain correct coro handles.
 - There seems to be a tiny binary size win.
 - A potential side benefit, which I'm currently not using, is that it makes `TaskPromise` (and analogs) open for inheritance without making them polymorphic classes. Right now, they are marked `final` in order to enable the non-polymorphic compiler optimizations. I'm not immediately pulling that `final`, since this creates real slicing risks, but the option value is nice.

Reviewed By: iahs

Differential Revision: D82905093

fbshipit-source-id: d3069e96385140b955c5cfd1c3dd5472a9b9e1ec