Expose more symbolizer functionality
Summary:
This diff changes the signature of:
 - `std::string getStackTraceStr();`
 - `std::string getAsyncStackTraceStr()`

Both now accept a `skip` param, which defaults to 0. `skip` omits the first `skip` frames from the stack trace string. This was already supported by the lower level printing code. This diff just exposes it.

The use case for this is logging a stacktrace in response to some in app event (e.g., an assert). For debugging purposes, the callstack should end at whatever triggered the need to print a stacktrace. Adding the various local and folly helper methods that actually generated the stacktrace to the printed output don't add value.

This diff also adds a new method:
```
template <size_t N>
FOLLY_ALWAYS_INLINE std::string getStackTraceStr(
    FrameArray<N>& fa, size_t skip = 0);
```

This allows getting a stacktrace as a string from previously captured frames. This allows for lazy symbolication.

Reviewed By: skrueger

Differential Revision: D83914416

fbshipit-source-id: 11c70bf5b0f92e10cdfd595de500788c8881366e