re-raise fatal signals with pidfd_send_signal on linux
Summary:
Using syscall `pidfd_send_signal` to the current thread, the fatal-signal handler can re-raise the current signal in the current thread with the current siginfo.

This is an alternative simply to returning from the signal handler and letting the faulting instruction re-execute, as was done in {D77041890}. The problem with re-executing is that the faulting instruction may not fault in its re-execution if the cause of the fault has concurrently been resolved.

Reviewed By: Jlalond

Differential Revision: D81533385

fbshipit-source-id: be7dd54a3985691d7adcd1f78753ee48a2684e66