folly/python/request_context rework Capsule Storage for Memory Leaks
Summary:
Ok I wrote a test to insure the use counts were not leaking.  But they were `test_multiple_ctx_copies_memory_leak` was ending up with a shared_ptr ref count of the number of tasks we spawned using each context.

This was all using shared_ptr::use_count so it wouldn't matter how many python objects were holding the capsule, we were leaking somehow, even with all the tasks/contextvar  deleted.

I wasn't able to track it down, but I replaced the PyCapsule with the existing Context object.    Context is really a PyCapsule with a easier to read API.

Reviewed By: itamaro, prakashgayasen

Differential Revision: D84285368

fbshipit-source-id: d4ba1f29ce1538d8e256144a649d4fc431bf4831