optimize the write path of MultiLevelTimeSeries
Summary:
As of today, values being added to a MultiLevelTimeSeries are cached and only flushed to the underlying BucketedTimeseries when the clock has moved to the next tick. The clock only moves forward on either writes (adding value) or reads (calling `update()`). This works fine until we changed the default clock type to be millisecond, meaning each tick become a lot shorter.

In this diff, we generalize the write path optimization in MultiLevelTimeSeries to cache the greatest-common-divisor of bucket sizes across all levels. E.g. A common MLTS has levels {60s, 600s, 0s /* all time */}, with a bucket count of 60. In this case, we will cache value in MLTS for 1s, despite the usage of a millisecond clock.

The correctness is ensured by preserving the invariant that if time point T belongs to bucket B, values added at T get reflected in bucket B. If readers are calling update() correctly, it follows that reading of an entire level (BucketedTimeSeries) or an interval of a level will observe exactly the same value as if every write were flushed immediately with no caching.

#buildall

Reviewed By: ot

Differential Revision: D83980306

fbshipit-source-id: 0e87fe48cb3aafe804361234da1644dae713ef8e