Rework Escape Hatch for folly::RequestContext::setContext watcher
Summary:
#buildmore
ig-fbcode-test

Since a pycontext watcher is always called before a setContext watcher is called,  we can store a thread local for the current pycontext,  which is on the python ThreadState but its not a public API and there exists no API to get it..    So if the setContext watcher fires and no PyContext is set then we can ignore the rest of the callback.

We can also simplify the PyState checks, since the only weird situation we need to know about is after finalization.

We grab the gil, because its safer now since we know that at some point this thread IS a python thread, and it protects from a potential use case where the GIL is dropped on a python thread and setContext is called.

Reviewed By: yfeldblum

Differential Revision: D84742790

fbshipit-source-id: 4db804e33e6bca6b6a9883572f8353697de609af