exception tracer: libc++ is only static in opt mode
Summary:
We made a mistake in D84536982. libc++ is statically linked only in opt build. The corresponding changes for exception tracer accidentally assumed libc++ is always statically linked in all libc++ build setups.

This caused registered callbacks to not always run.

This diff updated the condition for static linkage and made the libc++ static linking strategy specific to opt build.

Reviewed By: igorsugak

Differential Revision: D84839593

fbshipit-source-id: 7d028e73bb522cfa7ce0a26054d380a078e1ac71