Allow depending on third-party/liburing when using xlformers_llama4_mtia_conda
Summary:
This is required to allow using certain fbcode deps with the XLFormers MTIA conda stack.

This one has some nuances:

1. `third-party/liburing` is newer and has some additional symbols that are not present in the tp2 version.
2. We cannot default to using `liburing` from conda because that package is only available in `conda-forge` (many conda envs use `pytorch` and `main` channels instead).
3. We cannot redirect non-MTIA conda envs to `third-party/liburing` because that requires headers from a newer gcc version.

The opt-out for `liburing` in this target seems to work for other conda envs for now.

Since the MTIA conda env needs it (for the `cachelib/navy` dependency), this diff allows it to use the one from `third-party/liburing` which it can consume due to using a newer gcc version.

Reviewed By: bigfootjon

Differential Revision: D80558296

fbshipit-source-id: 1896490b169aa9f88536e3103de3a6cb9c77fb2a