Include generic iberty dep instead of iberty-demangle target
Summary:
D84776783 adds some changes that make `linux` the preferred dependency, however this overrides the previous `select()` that provides the `build-for-distro` constraint.

D37853390 introduces the `iberty-demangle` target, but this is problematic for us because this would include GPL-only code into our builds, and given that this target is `static`, this is strictly bad.

Instead, we should revert to our builds with `libiberty`, which is dynamically linked. That library is LGPL licensed, so dynamic linking is okay.

Reviewed By: igorsugak

Differential Revision: D84951262

fbshipit-source-id: 351cd3afffc3c18222a0aa0ec282fd9d4b857ca1