BasePromise de-duplicates task & async generator promises
Summary:
There'a bunch of shared functionality, primarily around `await_transform` user-observable behaviors, which are copy-pasted between `TaskPromise...` and `AsyncGeneratorPromise`.

Since I intend to add a bit more functionality here up-stack, I decided it was time to lift the duplication to a shared place.

This is definitely net-negative on "non-boilerplate lines". The only "new" piece of complexity I had to introduce was moving `AsyncGeneratorPromise::state_` into a templated `TailStorage tailStorage_` on the base. This was necessary to avoid making the promise 8 bytes bigger.

Reviewed By: ispeters

Differential Revision: D84993246

fbshipit-source-id: 9ceee862c100f43672dee77688994dd629eb50e3