Fix relay compiler build
Summary: In D84882493, the relay compiler was successfully built on Windows. However, D84936121 fixed a bug, but made relay windows build to fail once again. This diff addresses the build issue and also adds a CI configuration to ensure that the build is protected, as the CI currently only runs this build on Linux by default.

Reviewed By: lovexi

Differential Revision: D84972716

fbshipit-source-id: 05b97cb5407389a8e4b9b4c7b48b466a9bb7fab2