optimize folly::hazptr_domain hprec array
Summary:
Allocate the hprec's as an array - amortized, really a short chain of arrays.

A slight further optimization atop simply having a separate array for the hprecs walk in `load_hazptr_vals()` in `do_reclamation()`. Instead of walking a sequence of pointers to the hprec's, where the hprecs in the sequence are scattered about in memory, we walk a sequence of the hprecs themselves where the hprecs are mostly contiguous. Walking a sequence of pointers to scattered hprec's was already an optimization v.s. waling a linked-list of hprec's which were also scattered, since for very long lists with high scattering, the linked-list walk was stalled the execution pipeline and was very costly.

Reviewed By: paulmckrcu, DenisYaroshevskiy

Differential Revision: D83933374

fbshipit-source-id: ecf38582b751f6f827681bf832509d9225f77744