optimize list-walk to array-walk in hazptr_domain::load_hazptr_vals
Summary:
The code walks the linked list of `hazptr_rec` objects and inserts all attached `hazptr`'s into a `F14FastSet`.

The linked list may grow without bound, governed only by the maximum number of outstanding `hazptr`s - each live `hazptr` has to be attached to the list during its lifetime.

If there is some momentary usage spike, the list can grow large. But at steady state, the list nodes mostly do not have attached `hazptr`s.

In that scenario, the function walks the linked list periodically and, almost all of the time, observes `nullptr`; if not `nullptr`, inserts into the set.

A prior diff added the conditional to avoid inserting `nullptr`: {D82774448}. It is observed that prior to the diff, cycles were attributed to insertion into the set; after, cycles are not much moved, and are unattributed except to `load_hazptr_vals`.

The inference is that the linked list walk itself is pessimal, stalling the instruction pipeline. Which is, in general, a known phenomenon: walks over long linked lists in are pessimal.

Perhaps we can optimize this function by walking an array rather than walking a linked list. And we have the perfect array type for this scenario: `folly::atomic_grow_array`.

Note that this optimization leaves room. The hprecs in the array walk are still scattered about memory since the array walk is over an array of pointers to pointers to the hprec's, and the hprec's are still allocated separately. It would be possible to optimize further by allocating the hprec's mostly sequentially, but that would be a larger change.

Reviewed By: DenisYaroshevskiy

Differential Revision: D83933376

fbshipit-source-id: 178b75a5b3034822632b45337ff0a61fa15b9a1b