Fix compilation of F14Table with exceptions disabled
Summary: Fix the last remaining case of direct use of `try` / `catch` that may fail to compile with exceptions disabled. Unlike the previous ones fixed in D85084893 and D85385076, this one doesn't cause any actual errors at the moment but this can change in the future if `buildFromF14Table` gets used in a configuration with exceptions disabled.

Reviewed By: ilvokhin

Differential Revision: D85442051

fbshipit-source-id: 4f8ffe927b2b021de65250257503d58b369efc5b