allow all container and string types in split()
Summary:
Internal details of split() were restrictive of which container and string types were allowed. The purpose was to hide the template bodies in a single source file. But these sorts of restrictions can be more trouble than they are worth.

Uninstantiated templates are relatively cheap in terms of compile speed. Each individual instantiation of this is likely to be cheap. And there are likely to be few distinct instantiations per translation unit.

Reviewed By: DenisYaroshevskiy

Differential Revision: D85101083

fbshipit-source-id: 143201697af1e96e2ee787609dc5e402fa31fae2