Always update the firstTime_ field on both read and write path
Summary:
The current behavior of when `firstTime_` is updated is inconsistent between an all-time timeseries and non-all-time timeseries, and inconsistent between the write path (`addValue`) and the read path (`update()`). We take this opportunity to present a cleaner semantic of the `firstTime_` field and updated the documentation accordingly.

The spirit of the field is to capture when the timeseries started tracking (`update()` catures the read path, and `addValue` on the write path). So it can be used to calculate `elapsed` and `rate` by extension. It's most useful when the earliest trackable time (based on latestTime_) is older than `firstTime_`, so elapsed gets shortened properly.

With this diff we update the `firstTime_` field always (on all `update()` calls, not just the first one), and on all `addValue` calls as well, even when added value falls out of the tracked window. I believe the updated semantic is what has been intended all along, and it's a bug that it was implemented as such. The semantic change should have minimum impact on observed metrics for two reasons
1. for almost all meaningful use cases the timestamp is mostly moving forward
2. even if time does go back in time, for non-all-time tmeseries, the impact is limited to the beginning of the timeseries' lifetime, when the earliest trackable time is older than firstTime_ – it might produce a longer `elapsed` than before

Reviewed By: ot

Differential Revision: D85355998

fbshipit-source-id: 1bbdd90ce91921011fae865601ef7454dea3c9cb