expose `getEarliestTrackableTimeBy`
Summary:
In this diff, we add a const-qualified method `getEarliestTrackableTimeBy` that returns the earliest trackable time when the latest time is externally provided instead of using `latestTime_` always.

`getEarliestTrackableTimeBy` does not concern the actual aggregated stats. It will be used in later changes to enable `const`-qualified readers of the MultiLevelTimeSeries without having to call `update()` first.

Reviewed By: ot

Differential Revision: D85078954

fbshipit-source-id: 726fedc0ffb99466d8113a12b185f3a9e874e67a