add const-qualified methods for reading MultiLevelTimeSeries without calling `update()`
Summary:
This diff add `const`-qualified methods for reading MultiLevelTimeSeries without having to call `update()` first. Besides building on top of the `const` readers from the underlying BucketedTimeSeries, it still needs to handle value cached in the MultiLevelTimeSeries.

The diff introduces two helper methods that are doing the heavy-liftings. `totalBy` calculates the total (sum and count) for a given level, and also make a decision about if the cache value should be included in the final aggregate or not. For a given level, `elapsedBy`, takes everything into account (`cachedTime_`, `firstTime()`, `now`, `latestTime()`) and compute `elapsed()` as if `MultiLevelTimeSeries::update(now)` were called before reading the timeseries.

Reviewed By: ot

Differential Revision: D84937824

fbshipit-source-id: 06985e3c45795e3a661a0fda825c8cf5862070ac