still force libc++ to link as shared for now
Summary:
We can't force libc++ to be statically linked, even in opt mode due to the present of DSOs, which are widely used in Python environments.
Without this patch, cinderx wouldn't work due to its duplicate static and dynamic linkage to libc++.

While static linkage to libc++ remains possible in the absence of DSOs, let's force shared for now until we found a solution.

Reviewed By: skrueger

Differential Revision: D87374284

fbshipit-source-id: f8053c02c3cf3ba0d71247075103ff39b483c6cd