IoUringZeroCopyBufferPool: make sharable
Summary:
Make IoUringZeroCopyBufferPool sharable by moving the shared state into a shared_ptr private impl and making it into thin handle. The shared state has a struct io_uring as the *owning* ring. Each IoUringZeroCopyBufferPool handle is associated with its own struct io_uring, which may be different to the one in the shared state.

The initial ring that an IoUringZeroCopyBufferPool is created on becomes the owning ring. The only way to create IoUringZeroCopyBufferPool handles on other rings is to first *export* into an ExportHandle on the owning ring's EVB, move that handle to a new EVB, and then importing it.

The above process requires a new kernel API. The actual implementation will come in a later diff. No functional changes in this diff.

Reviewed By: vishwanath1306

Differential Revision: D87165860

fbshipit-source-id: 2c43066e9c541fa671251591e25861d7d185efc3